MY_RAND workaround update Message-Id: <011206114105.67cab@DUPHY4.Physics.Drexel.Edu>
p4raw-id: //depot/perl@13491

+++ vms/vms.c
@@ -7251,8 +7251,8 @@ Perl_sys_intern_clear(pTHX)
 void  
 Perl_sys_intern_init(pTHX)
 {
-    int ix = RAND_MAX;
-    float x;
+    unsigned int ix = RAND_MAX;
+    double x;
 
     VMSISH_HUSHED = 0;
