+++ vms/vmsish.h
@@ -298,7 +298,7 @@
 #define HAVE_INTERP_INTERN
 struct interp_intern {
     int    hushed;
-    float  inv_rand_max;
+    double inv_rand_max;
 };
 #define VMSISH_HUSHED     (PL_sys_intern.hushed)
 #define MY_INV_RAND_MAX   (PL_sys_intern.inv_rand_max)
